+++ app-ghc-9.2/GHCi/GhcApiCompat.hs
@@ -18,6 +18,7 @@ module GHCi.GhcApiCompat (
   , module GHC.Unit.Module.ModSummary
   , module GHC.Driver.Env.Types
   , module GHC.Runtime.Context
+  , module GHC.Types.FieldLabel 
   , CMC.catch
 ) where
 
@@ -40,4 +41,5 @@ import GHC.Types.TyThing
 import GHC.Unit.Module.ModSummary
 import GHC.Driver.Env.Types
 import GHC.Runtime.Context
+import GHC.Types.FieldLabel
 import qualified Control.Monad.Catch as CMC